+++ example/lib/widget/main_page.dart
@@ -1,24 +1,30 @@
 import 'package:example/widget/graph_area.dart';
 import 'package:example/widget/toolbar.dart';
-import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
+import 'package:plough/plough.dart';
 
 class MainPage extends StatefulWidget {
-  const MainPage({
-    super.key,
-  });
+  const MainPage({super.key});
 
   @override
   State<MainPage> createState() => _MainPageState();
 }
 
 class _MainPageState extends State<MainPage> {
+  final _viewportController = GraphViewportController();
+
+  @override
+  void dispose() {
+    _viewportController.dispose();
+    super.dispose();
+  }
+
   @override
   Widget build(BuildContext context) {
-    return const Column(
+    return Column(
       children: [
-        MainPageToolbar(),
-        Expanded(child: GraphArea()),
+        MainPageToolbar(viewportController: _viewportController),
+        Expanded(child: GraphArea(viewportController: _viewportController)),
       ],
     );
   }